ui/requirements.txt
Flask
mysql-connector-python 
requests